+++ eloh-app/components/SignInWithGoogleBtn.jsx
@@ -96,7 +96,7 @@ const GoogleSignInButton = ({ role }) => {
       <button
         onClick={handleSignIn}
         disabled={isLoading}
-        className="bg-[#03045e] text-white py-3 px-5 text-lg font-semibold rounded-xl shadow-[0_9px_#999] active:shadow-[0_5px_#666] active:translate-y-1 hover:bg-[#023e8a] transition-all duration-200 ease-in-out"
+        className="bg-[#03045e] text-white py-3 px-5 text-lg font-semibold rounded-xl shadow-[0_9px_#999] active:shadow-[0_5px_#666] active:translate-y-1 hover:bg-[#023e8a] transition-all duration-200 ease-in-out cursor-pointer"
       >
         {isLoading
           ? `${capitalizedRole} signing in...`
